requirements.txt
python-dotenv
model-context-protocol
uvicorn
aiohttp
